Change the mechanism of extension selection
The whole file list is now reselected each time the selection inside the
extension select changes. This is not the ideal option, and this should be
reworked so that selection only changes for files of the extensions for which
the selection has changed.

This could be easily done if there was an event for when selection changes for
a single option, but there isn't one. So the state of the extension selection
should be saved between selections and incorporated in file reselection